% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.AIC.r
\name{ch.AIC}
\alias{ch.AIC}
\title{A function calculate the AIC from a model fit}
\usage{
ch.AIC(y, fitY, numParameters, standardize = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{fitY}{A numeric vector containing the fit Y from the model.}

\item{numParameters}{The number of free parameters.}

\item{standardize}{A boolean that specifies whether to standardize the scores before calculating the AIC. Standardization is useful if you want to combine different DVs into a single AIC (e.g., RT and accuracy). DEFAULT = FALSE.
'}
}
\value{
the AIC.
}
\description{
This function calculates the AIC from a model fit.  The empirical Y and the fit Y, as well as the number of free parameters are input.  The AIC is output.  If one wants standardize Y and the fit Y before calculation, then set "standardize" to TRUE.
}
\examples{
ch.AIC (myY, fitY, 5)
}
\keyword{AIC}
