% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.altLogTransform.r
\name{ch.altLogTransform}
\alias{ch.altLogTransform}
\title{A function to log transform a variable that contains 0s and/or negatives}
\usage{
ch.altLogTransform(x)
}
\arguments{
\item{x}{a numeric combnVector.}
}
\value{
a vector of transformed numbers.
}
\description{
This function log transform a variable that contains 0s and/or negatives by first tranforming x == 0 into min(x[x>0])/2 and then transforming negative numbers into (min(x[x>0])/2)/abs(x).  Finally those numbers are log transformed.
}
\examples{
ch.altLogTransform (myX)
}
\keyword{alternative}
\keyword{log}
\keyword{negative}
\keyword{transform}
\keyword{zero}
