% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.altRootTransform.r
\name{ch.altRootTransform}
\alias{ch.altRootTransform}
\title{A function to transform a variable that contains 0s and/or negatives, while retaining ordinality, 0==0,
and symmetry of influence around 0}
\usage{
ch.altRootTransform(x, root = 4)
}
\arguments{
\item{x}{a numeric vector.}

\item{root}{a number indicating the root that x will be taken to (e.g., 2 = sqrt; 4=cubed root; etc.). DEFAULT = 4.}
}
\value{
a vector of transformed numbers.
}
\description{
This function transforms a variable as follows: x.out <- ifelse(x>0, x^(1/root), -1*abs(x)^(1/root)). This transformation compresses numbers that are a large distance from zero (both positive and negative).
}
\examples{
ch.altRootTransform (myX)
}
\keyword{alternative}
\keyword{negative}
\keyword{sqrt}
\keyword{transform}
\keyword{zero}
