% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.BIC.r
\name{ch.BIC}
\alias{ch.BIC}
\title{A function calculate the BIC from a model fit}
\usage{
ch.BIC(y, fitY, numParameters, standardize = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{fitY}{A numeric vector containing the fit Y from the model.}

\item{numParameters}{The number of free parameters.}

\item{standardize}{A boolean that specifies whether to standardize the scores before calculating the BIC. Standardization is useful if you want to combine different DVs into a single BIC (e.g., RT and accuracy). DEFAULT = FALSE.
'}
}
\value{
the BIC.
}
\description{
This function calculates the BIC from a model fit.  The empirical Y and the fit Y, as well as the number of free parameters are input.  The BIC is output.  If one wants standardize Y and the fit Y before calculation, then set "standardize" to TRUE.
}
\examples{
ch.BIC (myY, fitY, 5)
}
\keyword{BIC}
