% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binByNumElements.r
\name{ch.binByNumElements}
\alias{ch.binByNumElements}
\title{A function bins a vector of numbers}
\usage{
ch.binByNumElements(data, numBins)
}
\arguments{
\item{data}{A numeric vector.}

\item{numBins}{The number of bins you want the vector divided into.
'}
}
\value{
A vector of bins that match to the rows of "data".
}
\description{
This function creates N bins for a vector of numbers. It equalized the number of elements in each bin, so the intervals the each bin contains may vary based on the distribution of data. The bins are labeled by the average value of the items in the bin.
}
\examples{
ch.binByNumElements (rnorm(100, 10, 2), 10)
}
\keyword{bin}
\keyword{vector}
