% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.calculateDprimeStats.r
\name{ch.calculateDprimeStats}
\alias{ch.calculateDprimeStats}
\title{A function to calculate D Prime statistics}
\usage{
ch.calculateDprimeStats(
  data,
  grpCols,
  correctCol,
  correctVals = c(TRUE, FALSE),
  targetPresentCol,
  targetPresentVals = c(TRUE, FALSE),
  addCorrection = TRUE
)
}
\arguments{
\item{data}{a dataframe with the trial by trial responses.}

\item{grpCols}{a vector containing strings that specifies the names of the columns in "data" that you want the data grouped by.}

\item{correctCol}{a string that specifies the name of the column in "data" that contains the whether or not the participant responded correctly.}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{targetPresentCol}{a string that specifies the name of the column in "data" that contains the whether or not the target was presented.}

\item{targetPresentVals}{a vector of two values that specifies the "target present" value (index 1) and the "target absent" value (index 2). e.g, c("yes", "no")}

\item{addCorrection}{a boolean that specifies whether you want a .5 correction to be added the total hits, FAs, misses, and CRs. This corrects for 0 and 1 values for FA and Hits. DEFAULT = T.}
}
\value{
dataframe with the dprime statistics:
}
\description{
This function calculated D Prime statistics.
}
\examples{
ch.calculateDprimeStats (data=moralsData,"sn", "trial", "RT", "res.RT", "fit.RT", "overlap", "keyDef", c("Yes", "No"), "correct", params=parameters)
}
\keyword{analysis}
\keyword{dprime}
