% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.capwords}
\alias{ch.capwords}
\title{Capitalize first letter Function}
\usage{
ch.capwords(s, strict = FALSE)
}
\arguments{
\item{s}{a string of words.}

\item{strict}{a switch, if true, it ensures that only the first letter of each word is capitalized. Otherwise all remaining letters keep their default capitalization.  DEFAULT=FALSE.}
}
\value{
a string of words with the first letter capitalized
}
\description{
This function capitalizes the first letter of each word in a string of words.
}
\examples{
ch.capwords ("hello world")
}
\keyword{Capitalize}
