% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.extractDFbyGroups.r
\name{ch.extractDFbyGroups}
\alias{ch.extractDFbyGroups}
\title{A function to extract a subsetted dataframe returned the function ch.subsetDFbyGroups()}
\usage{
ch.extractDFbyGroups(subsettedDFlist, grpSubsetCond)
}
\arguments{
\item{subsettedDFlist}{the list of subsetted dataframes returned by ch.subsetDFbyGroups().}

\item{grpCols}{a single rowed dataframe that contains only the grouping variables used as an argument in ch.subsetDFbyGroups().  The first (and only) row should contain the value of each grouping variable that you want the dataframe to contain. For example, grpSubsetCond <- data.frame(gender = "Female", condition = "math").}
}
\value{
a dataframe that is identified by grpSubsetCond
}
\description{
This function extracts a specific dataframe the list of dataframes that is returned from the function ch.subsetDFbyGroups().  To identify the dataframe to be extracted, you create a one line dataframe that contains only the grouping variables used as an argument in ch.subsetDFbyGroups().  However, the first (and only) row should contain the value of each grouping variable that you want the dataframe to contain.  For example, grpSubsetCond <- data.frame(gender = "Female", condition = "math").
}
\examples{
ch.subsetDFbyGroups (outList, data.frame(gender = "Female", condition = "math"))
}
\keyword{dataframe}
\keyword{extract}
\keyword{grouping}
\keyword{subset}
\keyword{variables}
