% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.filterDataBetween.r
\name{ch.filterDataBetween}
\alias{ch.filterDataBetween}
\title{A function to filter data between two thresholds}
\usage{
ch.filterDataBetween(data, dataCol, lowThresh = NULL,
  highThresh = NULL)
}
\arguments{
\item{data}{a dataframe containing the data.}

\item{dataCol}{a string that specifies the name of the column in "data" on which the thresholds are to act.}

\item{lowThresh}{the lower threshold that the data must be above to be kept.}

\item{highThresh}{the higher threshold that the data must be below to be kept.}
}
\value{
a list containing: (datRemoved) a dataframe with the removed data; (datKept) a dataframe with the kept data; (numRemoved) the number of datapoints removed; (pRemoved) the proportion of datapoints removed.
}
\description{
This function filters the data between a low threshold and/or a high threshold. It returns a list that contains a dataframe of the removed data, a dataframe of the the kept data, the number of datapoints removed, and the percentage of points removed.
}
\examples{
ch.filterDataBetween (data=myDataFrame, "RT", 200, 5000)
}
\keyword{between}
\keyword{data}
\keyword{filter}
