% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.filterFastSubjects.r
\name{ch.filterFastSubjects}
\alias{ch.filterFastSubjects}
\title{A function to exclude participant whos RT is to fast}
\usage{
ch.filterFastSubjects(
  data,
  rtCol,
  subCol,
  trialCol = NULL,
  fast_ms = 250,
  trial_percent = 0.1,
  outputFileName = NULL,
  append = TRUE,
  plotFilename = NULL,
  removeFastResponses = FALSE
)
}
\arguments{
\item{data}{a dataframe with the trial by trial responses.}

\item{rtCol}{a string that specifies the name of the column in "data" that contains the RT data.}

\item{subCol}{a string that specifies the name of the column in "data" that contains the subject identifiers.}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number. It is used to plot the data.  DEFAULT = NULL (no plots)}

\item{fast_ms}{an integer identifying the ms threshold, equal to or below which are considered too fast. DEFAULT = 250}

\item{trial_percent}{a proportion (between 0-1) that identifies the proportion of trials below fast_ms that disqualify a participant. DEFAULT = 0.1}

\item{outputFileName}{a string specifying the output file name}

\item{append}{a boolean that specifies whether to append the ouput to the file if it exists. DEFAULT = TRUE.}

\item{plotFilename}{A string that identifies the name of file (.pdf) in which the plot will be saved. The default is NULL, whereby the plot will not be saved.}

\item{removeFastResponses}{a boolean that specifies whether to remove individual fast RT trials from final dataset. DEFAULT = FALSE.}
}
\value{
a list that contains datRemoved = the data that was removed; datKept = the data that was kept; numRemoved = the number of subjects removed; pRemoved = the percent of subjects removed
}
\description{
This function adds a legend to a plot.
}
\examples{
ch.filterFastSubjects(df.data, "rt", "sn")
}
\keyword{RT}
\keyword{fast}
\keyword{sn}
\keyword{subject}
