% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.filterDataBetween.r
\name{ch.filterGrpBtwn}
\alias{ch.filterGrpBtwn}
\title{A function to filter data between two thresholds grouped by a single column}
\usage{
ch.filterGrpBtwn(data, dataCol, grpCol, lowThresh = NULL,
  highThresh = NULL, FUN)
}
\arguments{
\item{data}{a dataframe containing the data.}

\item{dataCol}{a string that specifies the name of the column in "data" on which the thresholds are to act.}

\item{grpCol}{a string that specifies the name of the column in "data" that acts as a grouping variable.}

\item{lowThresh}{the lower threshold that the data must be above to be kept.}

\item{highThresh}{the higher threshold that the data must be below to be kept.}

\item{FUN}{the function that specifies the how to combine dataCol by grpCol (e.g., mean).}
}
\value{
a list containing: (datRemoved) a dataframe with the removed data as grouped by grpCol; (datKept) a dataframe with the kept data as grouped by grpCol; (numRemoved) the number of datapoints removed; (pRemoved) the proportion of datapoints removed; (datKeptRaw) a dataframe with the kept data in the raw format (ungrouped);.
}
\description{
This function filters the data, based on a grouping variable, between a low threshold and/or a high threshold. It returns a list that contains a dataframe of the removed data, a dataframe of the the kept data, the number of datapoints removed, and the percentage of points removed.
}
\examples{
ch.filterGrpBtwn (data=myDataFrame, "RT", "sn", 200, 5000, mean)
}
\keyword{between}
\keyword{data}
\keyword{filter}
\keyword{group}
