% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.get11diffPchs.r
\name{ch.get11diffPchs}
\alias{ch.get11diffPchs}
\title{A function to create 11 different pch symbols for plotting}
\usage{
ch.get11diffPchs()
}
\value{
a vector of pch values
}
\description{
This function creates 11 different pch symbols for plotting.
}
\examples{
ch.get11diffPchs()
}
\keyword{11}
\keyword{different}
\keyword{pch}
\keyword{plot}
