% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.get67diffLtys.r
\name{ch.get67diffLtys}
\alias{ch.get67diffLtys}
\title{A function to create 67 different line type values for plotting}
\usage{
ch.get67diffLtys()
}
\value{
a vector of lty values
}
\description{
This function creates 67 different line type values for plotting.
}
\examples{
ch.get70diffLtys()
}
\keyword{67}
\keyword{different}
\keyword{lty}
\keyword{plot}
