% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getChlogisticXfromProbability.r
\name{ch.getChlogisticXfromProbability}
\alias{ch.getChlogisticXfromProbability}
\title{A function to get the X value associated with a specified probability from ch.logistic}
\usage{
ch.getChlogisticXfromProbability(modelList, value = 0.5, range.x = NULL)
}
\arguments{
\item{modelList}{A list output by ch.logistic.}

\item{range.x}{A vector of two numbers, c(min, max), that specifies the range of x to look for pValue. If NULL or invalid, then the range will be the min and max of the x vector. DEFAULT = NULL}

\item{pValue}{A number between 0 and 1 that is the probability that you are interested in. DEFAULT = 0.5}
}
\value{
a number that is the X value associated with the input pValue.
}
\description{
This function gets the X value associated with a specified probability from ch.logistic
}
\examples{
ch.getChlogisticXfromProbability (modelList, 0.3)
}
\keyword{function}
\keyword{logistic}
\keyword{probability}
