% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getDemographics.r
\name{ch.getDemographics}
\alias{ch.getDemographics}
\title{This function calculates summary statistics for demographic data.}
\usage{
ch.getDemographics(
  data,
  subjectColumn,
  ageColumn = NULL,
  sexColumn = NULL,
  genderColumn = NULL,
  ethnicityColumn = NULL,
  outputFileName = "demographics.txt",
  append = F,
  roundValue = 2
)
}
\arguments{
\item{data}{A dataframe with the demographic information.  It needs - at minimum - a subject column and one other of the specified columns.}

\item{subjectColumn}{A string specifying the column of data that holds the subject identifier.}

\item{ageColumn}{A string specifying the column of data that holds the age information.  DEFAULT = NULL.}

\item{sexColumn}{A string specifying the column of data that holds the sex information.  DEFAULT = NULL.}

\item{genderColumn}{A string specifying the column of data that holds the gender information.  DEFAULT = NULL.}

\item{ethnicityColumn}{A string specifying the column of data that holds the ethnicity information.  DEFAULT = NULL.}

\item{outputFileName}{A sting specifying the name of the output file that will contain the summary statistics.  Default is "demographics.txt".}

\item{append}{A boolean specifying whether to append to end of an existing file or to overwrite any existing file with the same name.  DEFAULT = FALSE (overwrite).}

\item{roundValue}{An integer that specifies how many digits the output should be rounded to.  DEFAULT = 2.
'}
}
\value{
nothing - a text file is saved.
}
\description{
Function that calculates summary statistics for demographic data.
}
\examples{
ch.getDemographics (data, "sn", "age", genderColumn = "gender",outputFileName = "output.txt", append = T)
}
\keyword{demographics}
