% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getDprimeVariance.r
\name{ch.getDprimeVariance}
\alias{ch.getDprimeVariance}
\title{A function to get the variance from a d prime}
\usage{
ch.getDprimeVariance(NtargetPresent, NtargetAbsent, probFA, probHit)
}
\arguments{
\item{NtargetPresent}{the number of target present trials.}

\item{NtargetAbsent}{the number of target absent trials.}

\item{probFA}{the proportion of False Alarms.}

\item{probHit}{the proportions of Hits.}
}
\value{
the variance of the dPrime. Take the square root to get the SD.
}
\description{
This function computes the variance of a d prime using Miller (1996).
Miller, J. (1996). The sampling distribution of d’. Perception & Psychophysics, 58, 65-72.
much of the code taken from Suero, M., Privado, J., & Botella, J. (2017). Methods to estimate the variance of some indices of the signal detection theory: A simulation study. Psicológica, 38(1).
}
\examples{
ch.getDprimeVariance (100,100, .2, .7)
}
\keyword{dprime}
\keyword{variance}
