% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getGrpIndex.r
\name{ch.getGrpIndex}
\alias{ch.getGrpIndex}
\title{A function to get the df.grpIndex required to create legend values and a legend using ch.getPlotLegendVals () and ch.addLegend.}
\usage{
ch.getGrpIndex(data, grpCols)
}
\arguments{
\item{data}{the dataframe containing the grouped data.}

\item{grpCols}{a vector of strings specifying the column names of the grouping variables.}
}
\value{
a dataframe with the factorial combination of groups in data
}
\description{
This function creates a dataframe that contains the the factorial combination of grouping columns.
}
\examples{
ch.getGrpIndex (myDataframe, c("gender", "condition"))
}
\keyword{dataframe}
\keyword{df.grpIndex}
\keyword{grouping}
\keyword{subset}
\keyword{variables}
