% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getHSVincrement.r
\name{ch.getHSVincrement}
\alias{ch.getHSVincrement}
\title{A function to calculate the increment for HSV color based numChages needed}
\usage{
ch.getHSVincrement(numChanges, bounds = c(min, max))
}
\arguments{
\item{numChanges}{an integer denoting the number of changes required.}

\item{bounds}{the min and max of the h or s or v (usually 0 and 1).}
}
\value{
decimal indicating the hsv increment
}
\description{
This function calculates the increment for HSV color based numChages needed.
}
\examples{
ch.getHSVincrement(3, c(1,0))
}
\keyword{changes}
\keyword{groups}
\keyword{hsv}
\keyword{increment}
