% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getLearningEffect.r
\name{ch.getLearningEffect}
\alias{ch.getLearningEffect}
\title{A function to assess the learning effect in data}
\usage{
ch.getLearningEffect(data, x, y, a = 2000, b = -1, c = 100, ...)
}
\arguments{
\item{data}{the dataframe.}

\item{x}{a string the specifies the column name in "data" that contains the x variable.}

\item{y}{a string the specifies the column name in "data" that contains the y variable.}

\item{a}{a numeric value for the startpoint specifying the multiplier. DEFAULT = 2000.}

\item{b}{a numeric value for the startpoint specifying the multiplier in the exponential. DEFAULT = -1.}

\item{c}{a numeric value for the startpoint specifying the addend in the exponential. DEFAULT = 100.}
}
\value{
a nls fit function.
}
\description{
This function attempts to fit a non-linear, decelerating function to the specified data.
}
\examples{
ch.getLearningEffect (data=moralsData, trial, RT)
}
\keyword{data}
\keyword{morals}
\keyword{prep}
