% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getLmModel.r
\name{ch.getLmModel}
\alias{ch.getLmModel}
\title{a function to return an LM model that can be evaluated (i.e, an expression) from an lm object}
\usage{
ch.getLmModel(
  linearFit,
  yLab = NULL,
  xLab = expression(paste("", Psi, "(value) Distributional overlap", sep = ""))
)
}
\arguments{
\item{linearFit}{an lm object.}
}
\value{
expression with of the best fitting model (with parameters and predictor variables)
}
\description{
This function returns an LM model that can be evaluated (i.e, an expression) from an lm object
}
\examples{
\dontrun{
model <- ch.getLmModel(lmObject)
with(data, eval(model))
}
}
\keyword{eval}
\keyword{expression}
\keyword{lm}
\keyword{model}
\keyword{object}
