% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getPhitModel.r
\name{ch.getPhitModel}
\alias{ch.getPhitModel}
\title{a function to return a p(Hit) model that can be evaluated (i.e, an expression) from an nls object}
\usage{
ch.getPhitModel(
  pHitFit,
  yLab = "p(Hit)",
  xLab = expression(paste("", Psi, "(value) Distributional overlap", sep = ""))
)
}
\arguments{
\item{pHitFit}{an nls object - specifically from ch.pHVOfit().}
}
\value{
expression with of the best fitting model (with parameters and predictor variables)
}
\description{
This function returns a p(Hit) model that can be evaluated (i.e, an expression) from an nls object
}
\examples{
\dontrun{
model <- ch.getPhitModel(nlsObject)
with(data, eval(model))
}
}
\keyword{eval}
\keyword{expression}
\keyword{model}
\keyword{nls}
\keyword{object}
\keyword{p(Hit)}
