% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getPlotAxisMinMax.r
\name{ch.getPlotAxisMinMax}
\alias{ch.getPlotAxisMinMax}
\title{A function calculates the min and max values for a plot axis}
\usage{
ch.getPlotAxisMinMax(x, pBuffer = 0.1)
}
\arguments{
\item{variable}{a numeric array from which you the axis values are derived (usually just a vector).}

\item{p(buffer)}{the proportion of the axis range that you want to leave buffered on the low and the high end}
}
\value{
a vector containing the min and max values for the axis.
}
\description{
This function calculates the min and max values for a plot axis with a bit of a visual buffer.
}
\examples{
ch.getPlotAxisMinMax (xAxisVals)
}
\keyword{axis}
\keyword{buffer}
\keyword{plot}
