% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getPlotLegendVals.r
\name{ch.getPlotLegendVals}
\alias{ch.getPlotLegendVals}
\title{A function to create unique color and line type values for up to two groups to be plotted}
\usage{
ch.getPlotLegendVals(df.grpIndex)
}
\arguments{
\item{df.grpIndex}{a dataframe containing a column for each grouping variable with the values containing the combination of those variables that are to be plotted. The last column should be the index number for the model that will be used for plotting.  This dataframe is output by the XXX function.}

\item{groupsToPlot}{a vector of up to two strings containing the column names in df.grpIndex that the legend will be created for.}
}
\value{
.
}
\description{
This function creates unique color and line type values for groups to be plotted.
}
\examples{
ch.getPlotLegendVals(df.Index)
}
\keyword{colors}
\keyword{group}
\keyword{grouping}
\keyword{legend}
\keyword{lty}
\keyword{plot}
\keyword{variables}
