% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getPlotLegendValsFromData.r
\name{ch.getPlotLegendValsFromData}
\alias{ch.getPlotLegendValsFromData}
\title{A function to create unique color, line type, and point type for up to two groups to be plotted from a dataframe}
\usage{
ch.getPlotLegendValsFromData(
  data,
  grpCols,
  maxIntensityChanges = 8,
  maxHueChanges = 10
)
}
\arguments{
\item{data}{the dataframe to be plotted.}

\item{grpCols}{a vector of strings specifying the column names of the grouping variables.}
}
\value{
a list with all the subsetted dataframes
}
\description{
This function creates unique color, line type, and point type for up to two groups to be plotted from a dataframe
}
\examples{
ch.getPlotLegendValsFromData (myDataframe, c("gender", "condition"))
}
\keyword{dataframe}
\keyword{grouping}
\keyword{legend}
\keyword{variables}
