% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getProbMax.r
\name{ch.getProbMax}
\alias{ch.getProbMax}
\title{This function calculates probMax for chMaxAveComb based on the percent influence of the maximum valued item}
\usage{
ch.getProbMax(
  list.data,
  percentMaxInfluenceOverall = 0.6,
  percentMaxInfluenceOnAve = NULL
)
}
\arguments{
\item{list.data}{A list with each element filled with numbers. Each vactor in the list is a set of values for an item}

\item{percentMaxInfluenceOverall}{A probability that specifies the percent of influence that the column with the max(median()) has on the output vector over and above the average. When percentMaxInfluenceOverall = 0, then the max(median()) has no influence over and above the average. It's only influence is that it is part of the average.  When percentMaxInfluenceOverall > 0, then max(median()) has some influence on the average, but also some more influence over and above the average. When percentMaxInfluenceOverall = 1, then max(median()) has all the influence. pDEFAULT = 0.6 (this is the percent found with the biased average grouping model;}

\item{percentMaxInfluenceOnAve}{A probability that specifies the percent of influence that the column with the max(median()) has on the the average. When percentMaxInfluenceOnAve = 0, then the max(median()) has no influence the average.  When percentMaxInfluence > 0, then max(median()) has some influence on the average. When percentMaxInfluence = 1, then max(median()) has all the influence. pDEFAULT = NULL (this will force the column with the max(median()) to have an equal influence on the ave as all other columns);}
}
\value{
a probability to be input into ch.maxAveComb for the probMax option
}
\description{
This function calculates probMax for chMaxAveComb based on the percent influence of the maximum valued item
}
\examples{
ch.maxAveComb (list.data, percentMaxInfluence = .6)
}
\keyword{combine}
\keyword{function}
\keyword{vectors}
