% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getRowsByColItemsFiles.r
\name{ch.getRowsByColItemsFiles}
\alias{ch.getRowsByColItemsFiles}
\title{A function to extract the rows in FileA that contain the elements in FileB.}
\usage{
ch.getRowsByColItemsFiles(datafile, itemColName, colItemFile, saveFile = NULL)
}
\arguments{
\item{datafile}{a string specifying the name of the file containing the vector of elements used to identify the rows to extract from the datafile (i.e., FileB).}

\item{itemColName}{a string specifying the column name in the datafile that contains the items to be matched.}

\item{colItemFile}{a file containing a dataframe (i.e., FileA).}

\item{saveFile}{a string specifying the name of the file to save the output. DEFAULT = NULL (no output).}
}
\value{
a dataframe containing the extracted rows.
}
\description{
This function extracts the rows in FileA that contain the elements in FileB. FileA is a dataframe and FileB is a vector.
}
\examples{
ch.getRowsByColItemsFiles ("myfile.txt", "prompt", "myItems.txt", "myOutput.txt")
}
\keyword{dataframe}
\keyword{elements}
\keyword{extract}
\keyword{from}
\keyword{rows}
