% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.IC.r
\name{ch.IC}
\alias{ch.IC}
\title{A function calculate the AIC and BIC from a model fit}
\usage{
ch.IC(y, fitY, numParameters, standardize = FALSE, ICtype = "BIC")
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{fitY}{A numeric vector containing the fit Y from the model.}

\item{numParameters}{The number of free parameters.}

\item{standardize}{A boolean that specifies whether to standardize the scores before calculating the BIC. Standardization is useful if you want to combine different DVs into a single BIC (e.g., RT and accuracy). DEFAULT = FALSE.}

\item{ICtype}{A string specifying whether to return the AIC or BIC.  Valid inputs are: "AIC" and "BIC".  DEFAULT = "BIC"
'}
}
\value{
the AIC or BIC.
}
\description{
This function calculates the AIC and BIC from a from a model fit.  The empirical Y and the fit Y, as well as the number of free parameters are input.   If one wants standardize Y and the fit Y before calculation, then set "standardize" to TRUE. ICtype specifies whether to output the AIC or BIC.
}
\examples{
ch.IC (myY, fitY, 5, ICtype = "AIC")
}
\keyword{AIC}
\keyword{BIC}
\keyword{Criterion}
\keyword{IC}
\keyword{Information}
