% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.ICfromRSS.r
\name{ch.ICfromRSS}
\alias{ch.ICfromRSS}
\title{A function calculate the AIC and BIC from a residual sum of squares}
\usage{
ch.ICfromRSS(rss, n, numParameters, ICtype = "BIC")
}
\arguments{
\item{rss}{The residual sum of squares from a model fit.}

\item{n}{The number of observation from which the rss was calculated.}

\item{numParameters}{The number of free parameters.}

\item{ICtype}{A string specifying whether to return the AIC or BIC.  Valid inputs are: "AIC" & "BIC".  DEFAULT = "BIC"
'}
}
\value{
the AIC or BIC.
}
\description{
This function calculates the AIC and BIC from a residual sum of squares.  The residual sum of squares, the number of data points, as well as the number of free parameters are input. ICtype specifies whether to output the AIC or BIC.
}
\examples{
ch.ICfromRSS (myRss, 40, 5, ICtype = "BIC")
}
\keyword{AIC}
\keyword{BIC}
\keyword{Criterion}
\keyword{IC}
\keyword{Information}
