% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.mergeDataInDir}
\alias{ch.mergeDataInDir}
\title{Merge all the files in a directory}
\usage{
ch.mergeDataInDir(directoryName, outputFile, extension = ".dat",
  sep = "\\t", header = T, ...)
}
\arguments{
\item{directoryName}{this is a string that specifies the name of the directory containing the datafiles.}

\item{outputFile}{this is a string that specifies the name of the merged output file.}

\item{extension}{this is a string that specifies extensions of the files to be merged. DEFAULT = ".dat"}

\item{sep}{this is a string that specifies the how the columns in the files are separated. DEFAULT = tab}

\item{header}{this is a boolean that specifies whether column headers are present. DEFAULT = T}
}
\value{
nothing (a file is written)
}
\description{
This function merges all the files in a directory.
}
\examples{
ch.mergeDataInDir ("data", "out.txt")
}
\keyword{data}
\keyword{directory}
\keyword{merge}
