% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.newDir}
\alias{ch.newDir}
\title{New Subdrectory Function}
\usage{
ch.newDir(mainDir, subDir)
}
\arguments{
\item{mainDir}{the directory under which the subdirectory will exist.}

\item{subDir}{the new subdirectory.}
}
\description{
This function creates a new subdirectory and moves into it. If the subdirectory already exist, then it just moves into it.
}
\examples{
ch.newDir (currentDir, newDir)
}
\keyword{directory}
\keyword{new}
