% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.numberVectorToGrps.r
\name{ch.numberVectorToGrps}
\alias{ch.numberVectorToGrps}
\title{A function to bin quantities into categories.}
\usage{
ch.numberVectorToGrps(data, quantCol, catNameCol, quantValueCuts)
}
\arguments{
\item{data}{a dataframe.}

\item{quantCol}{a string the specifies the column name in "data" that contains the variable with the quantity.  This will be "p1Quant" or "p2Quant" if the items were processed with ch.moralsEquateItemStringsForNumberQuantDataset().}

\item{catNameCol}{a string the specifies the new column name in "data" that will be created to hold the new category names.}

\item{quantValueCuts}{a vector of numbers that specifies the cutoff values for each quantity.  The first category is <= to the first cutoff value; each successive category (c1, c2, etc) is c1 < X <= c2; the final category is greater than the last number in the vector.}
}
\value{
the dataframe (data) with the new category variable added.
}
\description{
This function bins quantities into categories.
}
\examples{
ch.numberVectorToGrps (itemAnalDat, "p1Quant", c(1,9,40))
}
