% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.permute}
\alias{ch.permute}
\title{Permutation Function}
\usage{
ch.permute(n)
}
\arguments{
\item{n}{the total number of items.}
}
\value{
a matrix of all the possible permutations of n items
}
\description{
This function creates all the permutations of all the items from 1-n
}
\examples{
ch.permute (4)
}
\keyword{permutations}
