% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.pHVOfit.r
\name{ch.pHVOfit}
\alias{ch.pHVOfit}
\title{Fits the p(HVO) function for the choice experiment}
\usage{
ch.pHVOfit(x, y, grp = NULL, useTwoParameterModel = FALSE)
}
\arguments{
\item{x}{the x variable for the x-axis (often overlap).}

\item{y}{the y variable for the y-axis (often p(hit)).}

\item{grp}{the grouping variable that identifies those stimuli whose value are above the reference distribution and those that are below the reference distribution. When this variable is included and useTwoParameterModel is set to TRUE, then the alpha parameter will be fit so it is symetric above and below 0.5 for refHVO and refLVO. DEFAULT = NULL (the grouping variable is ignored)}

\item{useTwoParameterModel}{A boolean that specifies whether to use a two parameter model.  If this is set to TRUE, then this function will fit a model whereby the rightmost point (overlap = 1.0) is not fixed at p(HVO) = 0.5. DEFAULT = FALSE.}
}
\value{
a list of the fit, r2 from the nls .
}
\description{
This function fits p(HVO) as a non-linear exponential decay function of an x variable (often overlap). It forces the first point to equal 1 and the last point is .5 on the y-axis. It outputs the fit of the function and the r_square.
}
\examples{
ch.pHVOfit (x,y, grp)
}
\keyword{fit}
\keyword{p(HVO)}
\keyword{p(hit)}
