% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.plot.lm}
\alias{ch.plot.lm}
\title{Graph function for Values Experiment}
\usage{
ch.plot.lm(x, y, plotTitle = NA, filename = NULL, cex1 = 1, printR2 = T,
  yLabel = NA, ylimMin = 0, ylimMax = 0, xlab = expression(paste("",
  Psi, "(value) Distributional overlap", sep = "")), ...)
}
\arguments{
\item{x}{the x variable for the x-axis (often overlap).}

\item{y}{the y variable for the y-axis (often p(hit)).}

\item{plotTitle}{a string with the title of the plot.}

\item{filename}{a string with the filename of the pdf of the file to be saved.  DEFAULT = NULL; no file saved.}

\item{cex1}{a numeric value for cex: the relative size of the text in the graph. cex1 > 1 is bigger; cex1 < 1 is smaller. DEFAULT=1.}

\item{printR2}{do you want the r square printed on the graph: TRUE/FALSE. DEFAULT=TRUE.}

\item{yLabel}{a string with the title of y-axis. DEFAULT=NA.}

\item{ylimMin}{a number denoting the minimum of the y-axis. DEFAULT=0.}

\item{ylimMax}{a number denoting the maximum of the y-axis. DEFAULT=0. If ylimMin == ylimMax, the function determines a pretty y-axis for you.}
}
\value{
the fit the lm .
}
\description{
This function creates a scatterplot (often a function of RT) as a function of an x variable (often overlap). It aldo fits a linear regression to the data. It outputs the fit of the function.
}
\examples{
ch.plot.lm (x,y)
}
\keyword{graph}
\keyword{overlap}
