% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.plot.pHit}
\alias{ch.plot.pHit}
\title{Graph function for Values Experiment}
\usage{
ch.plot.pHit(
  x,
  y,
  plotTitle = NA,
  filename = NULL,
  cex1 = 1,
  printR2 = T,
  yLabel = "p(hit)",
  ...
)
}
\arguments{
\item{x}{the x variable for the x-axis (often overlap).}

\item{y}{the y variable for the y-axis (often p(hit)).}

\item{plotTitle}{a string with the title of the plot.}

\item{filename}{a string with the filename of the pdf of the file to be saved.  DEFAULT = NULL; no file saved.}

\item{cex1}{a numeric value for cex: the relative size of the text in the graph. cex1 > 1 is bigger; cex1 < 1 is smaller. DEFAULT=1.}

\item{printR2}{do you want the r square printed on the graph: TRUE/FALSE. DEFAULT=TRUE.}

\item{yLabel}{a string with the title of y-axis. DEFAULT='p(hit)'}
}
\value{
a list of the fit, r2, and beta, from the nls .
}
\description{
This function creates a graph of p(hit) (0-1 on y-axis) as a function of an x variable (often overlap). It aldo fits a non-linear function to the data whereby the first point is 1 and the last point is .5. It outputs the fit of the function.
}
\examples{
ch.plot.pHit (x,y)
}
\keyword{graph}
\keyword{p(hit)}
