% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.plotTwoLinearFits.r
\name{ch.plotTwoLinearFits}
\alias{ch.plotTwoLinearFits}
\title{A function to fit and plot two linear function one over the other.}
\usage{
ch.plotTwoLinearFits(data, xCol, y1Col, y2Col, y1Label = NA, y2Label = NA,
  parOp = NULL, cex1 = 1.25, cex.topTile = 1.25, printR2 = T,
  topTitle = NULL, filename = NULL, ...)
}
\arguments{
\item{data}{a dataframe.}

\item{xCol}{a string that specifies the name of the column in "data" that contains the data for the x-axis.}

\item{y1Col}{a string that specifies the name of the column in "data" that contains the the first DV to be fit.}

\item{y2Col}{a string that specifies the name of the column in "data" that contains the the second DV to be fit.}

\item{y1Label}{a string that specifies the y-axis label name for the y1Fit. DEFAULT = NA.}

\item{y2Label}{a string that specifies the y-axis label name for the y2Fit. DEFAULT = NA.}

\item{parOp}{a list with the par options (see par). If NULL, then the functions default par options will be used. DEFAULT = NULL.}

\item{cex1}{sets the default font size. DEFAULT = 1.25.}

\item{cex.topTile}{sets the default font size of the title at the top of the page. DEFAULT = 1.25.}

\item{printR2}{a boolean that determines whether to print the r square on the graph. DEFAULT = FALSE.}

\item{topTitle}{a string that will be the title at the top of the page. DEFAULT = NULL.}

\item{filename}{the filename (pdf) to save the output of the graph. DEFAULT = NULL (no graph saved)}
}
\value{
a list containing: y1Fit = lm object with the fit of the y1 data; y2Fit = the lm  object with the fit of the y2 data.
}
\description{
This function fits  two linear functions;  plots the two functions on a single graph; and returns a list with the fit statistics.
}
\examples{
ch.moralsRTpHitFit (data=moralsData,"overlapRound", "resdRT", "correct", c("yes", "no"), filename = "plot.pdf")
}
\keyword{fit}
\keyword{linear}
\keyword{plots}
\keyword{two}
