% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.propVar.r
\name{ch.propVar}
\alias{ch.propVar}
\title{A function calculate the variance of a proportion}
\usage{
ch.propVar(p, n)
}
\arguments{
\item{p}{A numeric vector containing the proportions that variance will be calculated.}

\item{n}{A numeric vector containing the number of observations for each proportion.
'}
}
\value{
a numeric vector of the variances for each proportion in p.
}
\description{
This function calculates the variance of a proportion.
}
\examples{
ch.propVar (c(0.5, 0.2, 0.1), c(100, 150, 50))
}
\keyword{proportion}
\keyword{variance}
