% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.R2.r
\name{ch.R2}
\alias{ch.R2}
\title{A function calculate the r square from a model fit}
\usage{
ch.R2(y, fitY, standardize = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{fitY}{A numeric vector containing the fit Y from the model.}

\item{standardize}{A boolean that specifies whether to standardize the scores before calculating the BIC. Standardization is useful if you want to combine different DVs into a single BIC (e.g., RT and accuracy). DEFAULT = FALSE.
'}
}
\value{
the r square.
}
\description{
This function calculates the r2 from a from a model fit.  The empirical Y and the fit Y are input.   If one wants standardize Y and the fit Y before calculation, then set "standardize" to TRUE.
}
\examples{
ch.r2 (myY, fitY)
}
\keyword{r}
\keyword{r2}
\keyword{r_square}
\keyword{square}
