% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.removeBadSNs.r
\name{ch.removeBadSNs}
\alias{ch.removeBadSNs}
\title{A function to remove bad subjects' data from a dataframe}
\usage{
ch.removeBadSNs(data, snCol, badSNfilename)
}
\arguments{
\item{data}{a dataframe.}

\item{snCol}{a string that specifies the name of the column in "data" that contains the subject number.}

\item{badSNfilename}{the filename containing the IDs of the bad SNs.}
}
\value{
a dataframe without the bad subjects' data.
}
\description{
This function removes bad subjects' data from a dataframe. A filename containing the bad subject IDs are passed to the function.  This file should just be a vector (vertical in file) of IDs
}
\examples{
ch.removeBadSNs (data, "sn", "badSN.txt")
}
\keyword{bad}
\keyword{remove}
\keyword{subjects}
