% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.resize_jpg.r
\name{ch.resize_jpg}
\alias{ch.resize_jpg}
\title{A function to resize images.}
\usage{
ch.resize_jpg(
  input_file,
  output_file,
  width = NULL,
  height = NULL,
  keepAspectRatio = T
)
}
\arguments{
\item{input_file}{a string specifying the name of the input file containing the images.}

\item{output_file}{a string specifying the name of the output file that will contain the resized images.}

\item{width}{an integer specifying the new width of the image in pixels.}

\item{height}{an integer specifying the new height of the image in pixels.}

\item{keepAspectRatio}{A boolean specifying whether the orginal aspect ratio should be kept. If TRUE, then specify either the width or the height of the new image - NOT BOTH. DEFAULT = TRUE.}
}
\value{
Nothing.
}
\description{
This function resizes images.
}
\examples{
ch.resize_jpg ("inputImage.jpg", "outputImageFile.jpg", width = 200, keepAspectRatio=T)
}
\keyword{image}
\keyword{jpg}
\keyword{resize}
