% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.rgamma.r
\name{ch.rgamma}
\alias{ch.rgamma}
\title{A function to generate random numbers from a gamma distribution using the mean and sd as the input (rather than shape and scale)}
\usage{
ch.rgamma(n, mean, sd)
}
\arguments{
\item{n}{The number of observations.}

\item{m}{The mean of the gamma distribution.}

\item{s}{The standard deviation of the gamma distribution.
'}
}
\value{
A vector of observations from the requested gamma distribution.
}
\description{
This function to generate random numbers from a gamma distribution using the mean and sd as the input (rather than shape and scale)
}
\examples{
ch.rgamma (1000, 2, 0.5)
}
\keyword{distribution}
\keyword{gamma}
\keyword{random}
\keyword{rgamma}
