% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chUtils.r
\name{ch.round_any}
\alias{ch.round_any}
\title{New Rounding Function}
\usage{
ch.round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{a vector of numbers.}

\item{accuracy}{the accuracy of the rounding.}

\item{f}{the function for rounding: floor, ceiling, round.}
}
\description{
This function rounds numbers to a specified precision to in a particular direction.
}
\examples{
ch.round_any (x, .1, ceiling)
}
\keyword{any}
\keyword{numbers}
\keyword{round}
