% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.roundUpNice.r
\name{ch.roundUpNice}
\alias{ch.roundUpNice}
\title{A function to round pretty}
\usage{
ch.roundUpNice(x, nice = c(1, 2, 4, 5, 6, 8, 10))
}
\arguments{
\item{x}{the number to be rounded.}

\item{Nice}{a vector specifying the rounding parameters. The default values should work.}
}
\value{
the rounded number.
}
\description{
This function rounds up numbers pretty.
}
\examples{
cd.roundUpNice (6)
}
\keyword{Round}
\keyword{pretty}
\keyword{up}
