% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.RSS.r
\name{ch.RSS}
\alias{ch.RSS}
\title{A function calculate the residual sum of squares from a model fit}
\usage{
ch.RSS(y, fitY, standardize = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{fitY}{A numeric vector containing the fit Y from the model.}

\item{standardize}{A boolean that specifies whether to standardize the scores before calculating the RSS. Standardization is useful if you want to combine different DVs into a single RSS (e.g., RT and accuracy). DEFAULT = FALSE.
'}
}
\value{
the RSS.
}
\description{
This function calculates the residual sum of squares from a model fit.  The empirical Y and the fit Y are input.  The RSS is output.   If one wants to standardize the variables before calculation, then set "standardize" to TRUE.
}
\examples{
ch.RSS (myY, fitY)
}
\keyword{RSS}
\keyword{of}
\keyword{residual}
\keyword{squares}
\keyword{sum}
