% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.stringSeq.r
\name{ch.stringSeq}
\alias{ch.stringSeq}
\title{A function to create a sequence of string names}
\usage{
ch.stringSeq(stringName, seqVector)
}
\arguments{
\item{stringName}{a string that remains constant in the string names.}

\item{seqVector}{the elements in this vector are appended to the end of stringName.}
}
\value{
the vector of string names
}
\description{
This function creates a vector of string names with the elements in seqVector appended to the end of stringName. For example String1 String2 String3 ...
}
\examples{
ch.stringSeq ("string", c(1,2,3))
}
\keyword{append}
\keyword{sequence}
\keyword{string}
