% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.stripPlotByGrp.r
\name{ch.stripPlotByGrp}
\alias{ch.stripPlotByGrp}
\title{A function to plot a strip chart by groups.}
\usage{
ch.stripPlotByGrp(
  data,
  xCol,
  xSortColBy,
  grpCols = NULL,
  grpLgndNames = NULL,
  filename = NULL,
  lgndPlacement = "topright",
  cex1 = 1,
  cexLegend = 0.75,
  ylim = c(-1, 1),
  yLab = NA,
  pch = 1,
  las = 2,
  parOp = NULL,
  ...
)
}
\arguments{
\item{data}{a dataframe.}

\item{xCol}{a string the specifies the column name in "data" that contains the variable for the xColumn.}

\item{xSortColBy}{a string the specifies the column name in "data" that contains the variable for to sort the xColumn by.}

\item{grpCols}{a vector of strings the specifies the column names in "data" that contain the DVs for each group. The order is this vector matters, so sort them in the order you want them graphed (the colors will be relevant). This function depends on a "wide" format. So, each level of each group is in a separate column. DEFAULT = NULL}

\item{grpLgndNames}{a vector of strings the specifies the names you want each group to be called in the legend. The order is this vector matters, so sort them in the same order as grpCols.}

\item{filename}{the filename (pdf) to save the figure.  DEFAULT = NULL (file not saved)}

\item{lgndPlacement}{a string that specified the placement of the legend: "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". DEFAULT = "bottomright"}

\item{cex1}{the relative font size for the y-axis label. DEFAULT = 1.}

\item{cexLegend}{the relative font size for the legend. DEFAULT = .75.}

\item{ylim}{the limits of the y axis. DEFAULT = c(-1,1).}

\item{yLab}{y axis label. DEFAULT = NA.}

\item{pch}{see par for more info. DEFAULT = 1.}

\item{las}{see par for more info. DEFAULT = 2.}

\item{parOp}{the parameter list to enter into par.  DEFAULT = NULL}
}
\value{
.
}
\description{
This function plots a strip chart by groups.
}
\examples{
ch.stripPlotByGrp (overlapData, "probe", "meanDO", c("g1", "g2"))
}
