% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.subsetDFbyGroups.r
\name{ch.subsetDFbyGroups}
\alias{ch.subsetDFbyGroups}
\title{A function to subset a dataframe by the factorial combination of grouping columns}
\usage{
ch.subsetDFbyGroups(data, grpCols)
}
\arguments{
\item{data}{the dataframe to be subsetted.}

\item{grpCols}{a vector of strings specifying the column names of the grouping variables.}
}
\value{
a list with all the subsetted dataframes
}
\description{
This function subsets a dataframe by the factorial combination of grouping columns.  It outputs a list with each subsetted dataframe
}
\examples{
ch.subsetDFbyGroups (myDataframe, c("gender", "condition"))
}
\keyword{dataframe}
\keyword{grouping}
\keyword{subset}
\keyword{variables}
