% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.summariseBy.r
\name{ch.summariseBy}
\alias{ch.summariseBy}
\title{A function to summarize a dataframe one grouping variable}
\usage{
ch.summariseBy(data, grpCol, dvCol, newCol, FUN)
}
\arguments{
\item{data}{a dataframe.}

\item{grpCol}{a string specifying the name of the column in the dataframe containing the grouping variable.}

\item{dvCol}{a string specifying the name of the column in the dataframe containing the variable you want to summarize.}

\item{newCol}{a string specifying the name of the new column to be created that contains the summarized data.}

\item{FUN}{the summarizing function (e.g., mean).}
}
\value{
a dataframe
}
\description{
This function summarizes a dataframe by one grouping variable. It is easier to call than ddply within a function.
}
\examples{
ch.summariseBy (df, "sn", "RT", mean)
}
\keyword{ddply}
\keyword{group}
\keyword{summarize}
