% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.TSS.r
\name{ch.TSS}
\alias{ch.TSS}
\title{A function calculate the total sum of squares from a vector}
\usage{
ch.TSS(y, standardize = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{standardize}{A boolean that specifies whether to standardize the scores before calculating the RSS. Standardization is useful if you want to combine different DVs into a single RSS (e.g., RT and accuracy). DEFAULT = FALSE.
'}
}
\value{
the TSS.
}
\description{
This function calculates the total sum of squares from a vector.  The empirical Y is input.  The TSS is output.   If one wants to standardize the variable before calculation, then set "standardize" to TRUE.
}
\examples{
ch.TSS (myY)
}
\keyword{TSS}
\keyword{of}
\keyword{squares}
\keyword{sum}
\keyword{total}
