% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.updateRpackage.r
\name{ch.updateRpackage}
\alias{ch.updateRpackage}
\title{A function to update an R package}
\usage{
ch.updateRpackage(
  packageRoot = "~/Dropbox/UNCW/Active UNCW/research/Programming/R Packages",
  packageName = NULL
)
}
\arguments{
\item{packageRoot}{the directory in which the R package directory resides. DEFAULT = my computers packageRoot.}

\item{packageName}{the R package directory name.}
}
\value{
a list with all the parameterd
}
\description{
This function updates an R package easily.
}
\examples{
ch.readMoralsDBfile ("myDBfile.txt")
}
\keyword{dbfile}
\keyword{morals}
\keyword{read}
