% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.ZtestDprimeEqualZero.r
\name{ch.ZtestDprimeEqualZero}
\alias{ch.ZtestDprimeEqualZero}
\title{A function tests whether a dPrime is significantly Greater than/different from Zero}
\usage{
ch.ZtestDprimeEqualZero(NtargetPresent, NtargetAbsent, probFA, probHit,
  oneTailed = TRUE)
}
\arguments{
\item{NtargetPresent}{the number of target present trials.}

\item{NtargetAbsent}{the number of target absent trials.}

\item{probFA}{the proportion of False Alarms.}

\item{probHit}{the proportions of Hits.}

\item{oneTailed}{a boolean specifying whether the test is one tailed.  DEFAULT = TRUE (because we assume that dPrimes are bounded by Zero on the low end).}
}
\value{
a vector containing the z value (zVal) and the p value (p val) of the test.
}
\description{
This function tests whether a dPrime is significantly Greater than/different from Zero. It calculates variance using:
Miller, J. (1996). The sampling distribution of d’. Perception & Psychophysics, 58, 65-72.
}
\examples{
ch.ZtestDprimeEqualZero (100,100, .2, .7)
}
\keyword{different}
\keyword{dprime}
\keyword{test}
\keyword{z}
\keyword{zero}
