% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.r
\name{count}
\alias{count}
\title{A more versatile form of the T-SQL \code{count()} function.}
\usage{
count(..., condition = (function(x) TRUE))
}
\arguments{
\item{...}{an arbitrary number of \code{R} objects}

\item{condition}{a 1 argument condition}
}
\description{
Implementation of T-SQL \code{count} and Excel \code{COUNTIF} functions.
Shows the total number of elements in any number of data objects altogether or
that match a condition.
}
\details{
Originally written for the row.r package by Craig Varrichio. Included here because the rowr
package was discontinued. I use these functions in my packages
}
\examples{
count(c(NA,1,2))
count(c(NA,1,2),is.na)
count(c(NA,1,2),list('A',4),cbind(1,2,3))
count(c(NA,1,2),list('A',4),cbind(1,2,3),condition=is.character)
}
