% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.r
\name{len}
\alias{len}
\title{Allows finding the 'length' without knowledge of dimensionality.}
\usage{
len(data)
}
\arguments{
\item{data}{any \code{R} object}
}
\description{
Originally written for the row.r package by Craig Varrichio. Included here because the rowr
package was discontinued. I use these functions in my packages
}
\examples{
len(list(1,2,3))
len(c(1,2,3,4))
df<-data.frame(a=c(1,2,3),b=c(1,2,3))
len(df)
}
