% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.BIC.r
\name{ch.BIC}
\alias{ch.BIC}
\title{A function calculate the BIC from a model fit}
\usage{
ch.BIC(y, fitY, numParameters, standardize = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the empirical Y from the data.}

\item{fitY}{A numeric vector containing the fit Y from the model.}

\item{numParameters}{The number of free parameters.}

\item{standardize}{BIC is a function of the number of observations. If you wants correct for the number of observations, then set "standardize" to TRUE. DEFAULT = FALSE.
'}
}
\value{
the BIC.
}
\description{
This function calculates the BIC from a model fit.  The empirical Y and the fit Y, as well as the number of free parameters are input.  The BIC is output.  BIC is a function of the number of observations. If one wants correct for the number of observations, then set "standardize" to TRUE.
}
\examples{
ch.BIC (myY, fitY, 5)
}
\keyword{BIC}
