% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getPlotLegendVals.r
\name{ch.getPlotLegendVals}
\alias{ch.getPlotLegendVals}
\title{A function to create unique color and line type values for up to two groups to be plotted}
\usage{
ch.getPlotLegendVals(df.grpIndex)
}
\arguments{
\item{df.grpIndex}{a dataframe containing a column for each grouping variable with the values containing the combination of those variables that are to be plotted. The last column should be the index number for the model that will be used for plotting.  This dataframe is output by the ch.subsetDFbyGroups() function.}
}
\value{
this returns a dataframe containing the folloing columns: indexNum (the index number of the dataframe from ch.subsetDFbyGroups()); lty (the line type values for the graph); the group column names; h (the hue for hsv() color of line); s (the saturation for hsv() color of line);v (the intensity for hsv() color of line)
}
\description{
This function creates unique color and line type values for groups to be plotted.
}
\examples{
ch.getPlotLegendVals(df.Index)
}
\keyword{colors}
\keyword{group}
\keyword{grouping}
\keyword{legend}
\keyword{lty}
\keyword{plot}
\keyword{variables}
